requirements.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
joblib==0.16.0
lightgbm==2.3.1
numpy==1.19.0
pandas==1.0.5
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
sqlparse==0.3.1
threadpoolctl==2.1.0

# production requirements
gunicorn==20.0.4
awscli
